+++ test/kiln_cms_web/surface_test.exs
@@ -30,7 +30,8 @@ defmodule KilnCMSWeb.SurfaceTest do
     /editor/preview/:kind/:id /editor/redirects /editor/releases /editor/releases/:id
     /editor/search /editor/settings /editor/site-mail /editor/site-push /editor/site/:type/:slug /editor/slugs
     /editor/social
-    /editor/system /editor/tasks /editor/taxonomy /editor/team /editor/translations
+    /editor/site-storage /editor/system /editor/tasks /editor/taxonomy /editor/team
+    /editor/translations
     /editor/translations/export.xlf /editor/trash /editor/types /editor/webhooks /media
   )